refactor: 미사용 파일 제거 및 폴더 구조 변경 (#131)
