+++ src/pages/homePage/HomeHotEvent.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import EventCard from '../../components/events/EventCard';
+import EventCard from '../../components/EventCard';
 import { getPopularEvents } from '../../services/api/event';
 
 export default function HomeHotEvent() {
