+++ src/pages/searchPage/SearchEvent.jsx
@@ -1,4 +1,4 @@
-import EventCard from '../../components/events/EventCard';
+import EventCard from '../../components/EventCard';
 
 const SearchEvent = ({ events, bottomRef, loading, hasMore, isError }) => {
   return (
